+++ ext/polars/src/lib.rs
@@ -756,7 +756,7 @@ fn init(ruby: &Ruby) -> RbResult<()> {
     )?;
     class.define_method("with_context", method!(RbLazyFrame::with_context, 1))?;
     class.define_method("join_asof", method!(RbLazyFrame::join_asof, 12))?;
-    class.define_method("join", method!(RbLazyFrame::join, 9))?;
+    class.define_method("join", method!(RbLazyFrame::join, 10))?;
     class.define_method("with_column", method!(RbLazyFrame::with_column, 1))?;
     class.define_method("with_columns", method!(RbLazyFrame::with_columns, 1))?;
     class.define_method(